// generated/src/app/api-models/codereview-rules.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CodereviewRules {
  claimResultCode: string;
  processMethod: string;
  usePromatch: string;
  pmatchKbId: number;
  histChangeHoldReasn: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  origLineDenyReasn: string;
}